feat(web): add `update-label-button` for refactory `edit-panel`
