+++ web/package.json
@@ -55,6 +55,7 @@
     "eslint": "^9",
     "eslint-config-next": "^15.3.0",
     "eslint-config-prettier": "^10.1.1",
+    "eslint-import-resolver-typescript": "^4.3.4",
     "eslint-plugin-prettier": "^5.2.3",
     "eslint-plugin-react-hooks": "^5.2.0",
     "prettier": "^3.5.3",
@@ -83,6 +84,7 @@
     "class-variance-authority": "^0.7.1",
     "eslint-config-next": "^15.3.0",
     "eslint-config-prettier": "^10.1.1",
+    "eslint-import-resolver-typescript": "^4.3.4",
     "eslint-plugin-prettier": "^5.2.3",
     "eslint-plugin-react-hooks": "^5.2.0",
     "prettier-plugin-tailwindcss": "^0.6.11",
