+++ web/src/app/dashboard/workspace/[workspace]/@modal/(.)task/[task]/_components/task-add-label/provider.tsx
@@ -64,7 +64,7 @@ export default function TaskAddLabelProvider({
       editTag,
       setEditTag,
     }),
-    [open, view, editTag],
+    [open, setOpen, view, setView, editTag, setEditTag],
   );
 
   return (
